# lib/observability/http-router.js
'use strict';

const { readExactWorkspace } = require('../http/exact-workspace');

const PREFIX = '/api/observability';
const VERSIONED_PREFIX = '/api/v1/observability';
const API_PREFIXES = Object.freeze([VERSIONED_PREFIX, PREFIX]);
const NO_STORE = Object.freeze({
  'Cache-Control': 'no-store',
  'X-Content-Type-Options': 'nosniff',
});
const STREAM_HEADERS = Object.freeze({
  'Content-Type': 'text/event-stream; charset=utf-8',
  'Cache-Control': 'no-store',
  'Connection': 'keep-alive',
  'X-Accel-Buffering': 'no',
  'X-Content-Type-Options': 'nosniff',
});

function send(writeJson, req, res, status, body) {
  writeJson(req, res, status, body, NO_STORE);
}

function failure(writeJson, req, res, status, code, message, headers = NO_STORE) {
  writeJson(req, res, status, { ok: false, error: { code, message } }, headers);
}

function validateQuery(reqUrl) {
  const allowed = new Set(['workspaceId', 'limit', 'cursor', 'eventType', 'runId', 'status', 'windowMs']);
  if (reqUrl.search.length > 2_048 || [...reqUrl.searchParams.keys()].length > 8) return false;
  for (const key of reqUrl.searchParams.keys()) {
    const values = reqUrl.searchParams.getAll(key);
    if (key === 'workspaceId') continue;
    if (!allowed.has(key) || values.length !== 1 || values[0].length > (key === 'cursor' ? 512 : 128)) return false;
  }
  const limit = reqUrl.searchParams.get('limit');
  if (limit !== null && (!/^[1-9]\d*$/.test(limit) || Number(limit) > 100)) return false;
  const windowMs = reqUrl.searchParams.get('windowMs');
  return windowMs === null || (/^[1-9]\d*$/.test(windowMs) && Number(windowMs) >= 1_000 && Number(windowMs) <= 2_678_400_000);
}

function queryOptions(reqUrl) {
  return {
    limit: reqUrl.searchParams.get('limit') || undefined,
    cursor: reqUrl.searchParams.get('cursor') || undefined,
    eventType: reqUrl.searchParams.get('eventType') || undefined,
    runId: reqUrl.searchParams.get('runId') || undefined,
    status: reqUrl.searchParams.get('status') || undefined,
    windowMs: reqUrl.searchParams.get('windowMs') || undefined,
  };
}

function exactBodyWorkspace(body) {
  if (!body || typeof body !== 'object' || Array.isArray(body)
      || !Object.prototype.hasOwnProperty.call(body, 'workspaceId')) {
    return { ok: false, code: 'MISSING_WORKSPACE_ID' };
  }
  const workspaceId = typeof body.workspaceId === 'string' ? body.workspaceId.trim() : '';
  if (!workspaceId || workspaceId !== body.workspaceId || workspaceId.length > 128 || /[\x00-\x1F\x7F]/.test(workspaceId)) {
    return { ok: false, code: 'INVALID_WORKSPACE_ID' };
  }
  return { ok: true, workspaceId };
}

function matchApiPrefix(pathname) {
  for (const prefix of API_PREFIXES) {
    if (pathname === prefix || pathname.startsWith(`${prefix}/`)) return prefix;
  }
  return null;
}

function createObservabilityHttpRouter({ getService, parseJsonRequest, writeJson, denyIfUnauthorized, authorizeWorkspace, rateLimiter }) {
  for (const dependency of [getService, parseJsonRequest, writeJson, denyIfUnauthorized, authorizeWorkspace]) {
    if (typeof dependency !== 'function') throw new TypeError('observability route dependencies are required');
  }
  if (!rateLimiter || typeof rateLimiter.check !== 'function' || typeof rateLimiter.acquireStream !== 'function') {
    throw new TypeError('observability rate limiter is required');
  }

  function authorize(req, res, workspaceId, permission) {
    try {
      const decision = authorizeWorkspace({ principal: req.huqanAuth, workspaceId, permission });
      if (decision?.allowed === true) return true;
      failure(writeJson, req, res, 403, decision?.code || 'OBSERVABILITY_WORKSPACE_FORBIDDEN', 'Observability access is forbidden.');
    } catch (_) {
      failure(writeJson, req, res, 503, 'OBSERVABILITY_AUTHORIZATION_UNAVAILABLE', 'Observability authorization is unavailable.');
    }
    return false;
  }

  function limit(req, res, workspaceId, action) {
    try {
      const decision = rateLimiter.check({ principal: req.huqanAuth, workspaceId, action });
      if (decision.allowed) return true;
      failure(writeJson, req, res, 429, decision.code, 'Observability request limit exceeded.', {
        ...NO_STORE, 'Retry-After': String(decision.retryAfter || 1),
      });
    } catch (_) {
      failure(writeJson, req, res, 503, 'OBSERVABILITY_RATE_LIMIT_UNAVAILABLE', 'Observability rate limiting is unavailable.');
    }
    return false;
  }

  function acquireStream(req, res, workspaceId) {
    try { return rateLimiter.acquireStream({ principal: req.huqanAuth, workspaceId }); }
    catch (_) {
      failure(writeJson, req, res, 503, 'OBSERVABILITY_RATE_LIMIT_UNAVAILABLE', 'Observability rate limiting is unavailable.');
      return null;
    }
  }

  return async function handleObservabilityRoute(req, res, reqUrl) {
    const prefix = matchApiPrefix(reqUrl.pathname);
    if (!prefix) return false;
    const versioned = prefix === VERSIONED_PREFIX;
    const suffix = reqUrl.pathname.slice(prefix.length).replace(/\/$/, '') || '/';
    if (suffix === '/stream') {
      if (req.method !== 'GET') {
        failure(writeJson, req, res, 405, 'METHOD_NOT_ALLOWED', 'Method not allowed');
        return true;
      }
      if (!denyIfUnauthorized(req, res, NO_STORE)) return true;
      if (!validateQuery(reqUrl)) {
        failure(writeJson, req, res, 400, 'OBSERVABILITY_QUERY_INVALID', 'Observability query is invalid.');
        return true;
      }
      const workspace = readExactWorkspace(reqUrl.searchParams);
      if (!workspace.ok) {
        failure(writeJson, req, res, 400, workspace.code, 'Exactly one non-empty workspaceId is required.');
        return true;
      }
      if (!authorize(req, res, workspace.workspaceId, 'stream')) return true;
      if (!limit(req, res, workspace.workspaceId, 'streamAttempts')) return true;
      const streamLease = acquireStream(req, res, workspace.workspaceId);
      if (!streamLease) return true;
      if (!streamLease.allowed) {
        failure(writeJson, req, res, 429, streamLease.code, 'Observability stream limit exceeded.', {
          ...NO_STORE, 'Retry-After': String(streamLease.retryAfter || 1),
        });
        return true;
      }
      try {
        const service = getService();
        res.writeHead(200, STREAM_HEADERS);
        res.write(`event: ready\ndata: ${JSON.stringify({ ok: true, workspaceId: workspace.workspaceId })}\n\n`);
        const unsubscribe = service.subscribe(event => {
          if (event.workspaceId !== workspace.workspaceId || res.writableEnded) return;
          res.write(`event: ${event.eventType}\ndata: ${JSON.stringify(event)}\n\n`);
        });
        const keepAlive = setInterval(() => {
          if (!res.writableEnded) res.write(': keep-alive\n\n');
        }, 25_000);
        keepAlive.unref?.();
        const close = () => {
          clearInterval(keepAlive);
          unsubscribe();
          streamLease.release();
        };
        req.once('close', close);
        res.once('close', close);
      } catch (error) {
        streamLease.release();
        if (!res.headersSent) failure(writeJson, req, res, 503, error.code || 'OBSERVABILITY_UNAVAILABLE', 'Observability service is unavailable.');
      }
      return true;
    }

    if (!['/metrics', '/events', '/runs', '/queue', '/alerts', '/alert-rules'].includes(suffix)
        && !/^\/alert-rules\/[^/]+$/.test(suffix) && !/^\/alerts\/[^/]+\/acknowledge$/.test(suffix)) return false;
    if (!denyIfUnauthorized(req, res, NO_STORE)) return true;
    if (!validateQuery(reqUrl)) {
      failure(writeJson, req, res, 400, 'OBSERVABILITY_QUERY_INVALID', 'Observability query is invalid.');
      return true;
    }

    // DELETE reads its scope from the query string exactly as the GET surfaces
    // do, so it goes through the same validation. Without it a request that
    // names no workspace was rejected on GET but accepted on DELETE, where
    // normalizeWorkspaceId silently resolved the missing value to `default`
    // and removed a rule the caller never scoped.
    const requiresWorkspaceQuery = req.method === 'GET' || req.method === 'DELETE';
    let workspaceId = '';
    if (requiresWorkspaceQuery) {
      const workspace = readExactWorkspace(reqUrl.searchParams);
      if (!workspace.ok) {
        failure(writeJson, req, res, 400, workspace.code, 'Exactly one non-empty workspaceId is required.');
        return true;
      }
      workspaceId = workspace.workspaceId;
      const permission = req.method === 'DELETE' ? 'alerts:write' : 'read';
      if (!authorize(req, res, workspaceId, permission)) return true;
      const action = req.method === 'DELETE' ? 'alertMutations' : 'listRequests';
      if (!limit(req, res, workspaceId, action)) return true;
    }

    try {
      let body = null;
      if (req.method === 'POST' && suffix === '/queue') {
        body = await parseJsonRequest(req, res, { maxBytes: 12_288 });
        if (!body) return true;
        const workspace = exactBodyWorkspace(body);
        if (!workspace.ok) {
          failure(writeJson, req, res, 400, workspace.code, 'Exactly one non-empty workspaceId is required.');
          return true;
        }
        workspaceId = workspace.workspaceId;
        if (!authorize(req, res, workspaceId, 'queue:write')) return true;
        if (!limit(req, res, workspaceId, 'queueMutations')) return true;
      } else if (req.method === 'POST' && suffix === '/alert-rules') {
        body = await parseJsonRequest(req, res, { maxBytes: 4_096 });
        if (!body) return true;
        const workspace = exactBodyWorkspace(body);
        if (!workspace.ok) {
          failure(writeJson, req, res, 400, workspace.code, 'Exactly one non-empty workspaceId is required.');
          return true;
        }
        workspaceId = workspace.workspaceId;
        if (!authorize(req, res, workspaceId, 'alerts:write')) return true;
        if (!limit(req, res, workspaceId, 'alertMutations')) return true;
      } else if (req.method === 'POST' && /^\/alerts\/[^/]+\/acknowledge$/.test(suffix)) {
        body = await parseJsonRequest(req, res, { maxBytes: 2_048 });
        if (!body) return true;
        const workspace = exactBodyWorkspace(body);
        if (!workspace.ok) {
          failure(writeJson, req, res, 400, workspace.code, 'Exactly one non-empty workspaceId is required.');
          return true;
        }
        workspaceId = workspace.workspaceId;
        if (!authorize(req, res, workspaceId, 'alerts:write')) return true;
        if (!limit(req, res, workspaceId, 'alertMutations')) return true;
      }
      const service = getService();
      if (req.method === 'GET' && suffix === '/metrics') {
        send(writeJson, req, res, 200, { ok: true, data: { metrics: service.summary({ workspaceId, windowMs: reqUrl.searchParams.get('windowMs') }), queue: service.queueSummary({ workspaceId }), alerts: service.listAlerts({ workspaceId, limit: 20 }) } });
        return true;
      }
      if (req.method === 'GET' && suffix === '/events') {
        send(writeJson, req, res, 200, { ok: true, data: service.listEvents({ workspaceId, ...queryOptions(reqUrl) }) });
        return true;
      }
      if (req.method === 'GET' && suffix === '/runs') {
        send(writeJson, req, res, 200, { ok: true, data: service.listRuns({ workspaceId, ...queryOptions(reqUrl) }) });
        return true;
      }
      if (req.method === 'GET' && suffix === '/queue') {
        const items = versioned
          ? service.pageQueue({ workspaceId, ...queryOptions(reqUrl) })
          : { items: service.listQueue({ workspaceId, limit: reqUrl.searchParams.get('limit') }) };
        send(writeJson, req, res, 200, { ok: true, data: { ...service.queueSummary({ workspaceId }), ...items } });
        return true;
      }
      if (req.method === 'GET' && suffix === '/alerts') {
        const data = versioned
          ? service.pageAlerts({ workspaceId, ...queryOptions(reqUrl) })
          : { items: service.listAlerts({ workspaceId, limit: reqUrl.searchParams.get('limit') }) };
        send(writeJson, req, res, 200, { ok: true, data });
        return true;
      }
      if (req.method === 'GET' && suffix === '/alert-rules') {
        const data = versioned
          ? service.pageAlertRules({ workspaceId, ...queryOptions(reqUrl) })
          : { items: service.listAlertRules({ workspaceId, limit: reqUrl.searchParams.get('limit') }) };
        send(writeJson, req, res, 200, { ok: true, data });
        return true;
      }
      if (req.method === 'POST' && suffix === '/queue') {
        const job = service.enqueueJob(body);
        send(writeJson, req, res, 202, { ok: true, data: job });
        return true;
      }
      if (req.method === 'POST' && suffix === '/alert-rules') {
        const rule = service.createAlertRule(body);
        send(writeJson, req, res, 201, { ok: true, data: rule });
        return true;
      }
      const acknowledgeMatch = suffix.match(/^\/alerts\/([^/]+)\/acknowledge$/);
      if (acknowledgeMatch && req.method === 'POST') {
        const alert = service.acknowledgeAlert({ workspaceId, alertId: decodeURIComponent(acknowledgeMatch[1]) });
        if (!alert && versioned) {
          failure(writeJson, req, res, 404, 'OBSERVABILITY_ALERT_NOT_FOUND', 'Observability alert was not found.');
          return true;
        }
        send(writeJson, req, res, alert ? 200 : 404, { ok: Boolean(alert), data: alert });
        return true;
      }
      const ruleMatch = suffix.match(/^\/alert-rules\/([^/]+)$/);
      if (ruleMatch && req.method === 'DELETE') {
        const deleted = service.deleteAlertRule({ workspaceId, ruleId: decodeURIComponent(ruleMatch[1]) });
        if (!deleted && versioned) {
          failure(writeJson, req, res, 404, 'OBSERVABILITY_ALERT_RULE_NOT_FOUND', 'Observability alert rule was not found.');
          return true;
        }
        send(writeJson, req, res, deleted ? 200 : 404, { ok: deleted, data: { deleted } });
        return true;
      }
      failure(writeJson, req, res, 405, 'METHOD_NOT_ALLOWED', 'Method not allowed');
    } catch (error) {
      const status = /^INVALID_|^OBSERVABILITY_/.test(error.code || '') ? 400 : 503;
      failure(writeJson, req, res, status, error.code || 'OBSERVABILITY_FAILED', error.message || 'Observability request failed.');
    }
    return true;
  };
}

module.exports = { API_PREFIXES, createObservabilityHttpRouter, matchApiPrefix, NO_STORE, PREFIX, STREAM_HEADERS, VERSIONED_PREFIX };